Completely refactor names, places, tags as derived classes
There is now a PlornAttr base class that contains all the common
features of names, places, and tags.  PlornName et al are now
derived from that class allowing a lot of duplicate code to be
removed.

This simplifies the db code enormously and is intended to simplify
the GUI code later on.

Signed-off-by: Al Stone <ahs3@ahs3.net>